Return bech32 error for incorrect address string
When a decoding error occurs for a bech32 address string the error is
ignored and the same address string is attempted to be decoded as
base58.  This then incorrectly returns a base58 error.

This has been changed so that the prefix of the address string is
checked when there is a bech32 error, if it is a recognized bech32
prefix the error is returned.