Fix rustdoc example
The rustdoc example was designed to demonstrate a case that panics but
was marked with `compile_fail`.  The example code failed to compile because of
`error[E0658]: use of unstable library feature slice_as_chunks` so the
`compile_fail` erroneously allowed the crate to compile.

Add the feature gate suggested by the compiler.

Change the rustdoc tag to `should_panic`