Use AsRef<Network> instead of AsRef<Params>
Done as part of crate smashing. A few times we use `AsRef<Params>` as
a function parameter only to get the network field from it.

We can use `AsRef<Network>` instead. If we also implement
`AsRef<Network> for Params` then users can still pass in a `Params`
struct.

I'm not 100% sure this is correct as far as API design goes.